# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/advertisers" : {
      "get" : {
        "tags" : [ "advertiser-controller" ],
        "operationId" : "getAdvertiserByExternalRef",
        "parameters" : [ {
          "name" : "externalRef",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAdvertiserResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "advertiser-controller" ],
        "operationId" : "register",
        "parameters" : [ {
          "name" : "X-User-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AdvertiserRegistrationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAdvertiserResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/advertisers/{id}" : {
      "get" : {
        "tags" : [ "advertiser-controller" ],
        "operationId" : "getAdvertiser",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAdvertiserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/advertisers/{advertiserId}/presigned-url" : {
      "get" : {
        "tags" : [ "advertiser-controller" ],
        "operationId" : "getPresignedUrlForUpload",
        "parameters" : [ {
          "name" : "advertiserId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "fileName",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "contentType",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/advertisers/me" : {
      "get" : {
        "tags" : [ "advertiser-controller" ],
        "operationId" : "getMyAdvertiser",
        "parameters" : [ {
          "name" : "X-User-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAdvertiserResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "AdvertiserRegistrationRequest" : {
        "required" : [ "companyName" ],
        "type" : "object",
        "properties" : {
          "companyName" : {
            "type" : "string"
          },
          "externalRef" : {
            "type" : "string"
          }
        }
      },
      "AdvertiserResponse" : {
        "required" : [ "companyName", "createdAt", "id", "updatedAt", "userId" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "userId" : {
            "type" : "string"
          },
          "companyName" : {
            "type" : "string"
          },
          "externalRef" : {
            "type" : "string"
          },
          "walletBalanceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ApiResponseAdvertiserResponse" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AdvertiserResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ApiResponseString" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    }
  }
}